+++ tests/testthat/test_summarise.R
@@ -23,9 +23,12 @@ test_that("Summarises area by component",{
                         gifa = 1,
                         unit_area = c(1, 1, 2, 10, 20, 30))
 
-  correct <- data.frame(grade = c("A", "B", "D", "C", "E", "A", "B", "C", "D", "E"),
-                        elementid = c(1, 2, 2, 3, 3, 4, 4, 4, 4, 4),
-                        area = c(1, 6, 5, 2, 20, 12, 9, 6, 2.1, 0.9),
+  correct <- data.frame(grade = rep(c("A", "B", "C", "D", "E"), 4),
+                        elementid = rep(1:4, each = 5),
+                        area = c(1, 0, 0, 0, 0,
+                                 0, 6, 0, 5, 0,
+                                 0, 0, 2, 0, 20,
+                                 12, 9, 6, 2.1, 0.9),
                         stringsAsFactors = FALSE)
 
   expect_equal(as.data.frame(element_summarise_area(element, elementid)), # need to coerce result into data frame as expect equal tolerances don't work with tibbles.
@@ -55,9 +58,12 @@ test_that("Summarises backlog by component",{
                         gifa = 1,
                         unit_area = c(1, 1, 2, 10, 20, 30))
 
-  correct <- data.frame(grade = c("B", "D", "C", "E", "B", "C", "D", "E"),
-                        elementid = c(2, 2, 3, 3, 4, 4, 4, 4),
-                        backlog = c(6, 20, 6, 140, 9, 24, 1.68, 9),
+  correct <- data.frame(grade = rep(c("B", "C", "D", "E"), 4),
+                        elementid = rep(1:4, each = 4),
+                        backlog = c(0, 0, 0, 0,
+                                    6, 0, 20, 0,
+                                    0, 6, 0, 140,
+                                    9, 24, 1.68, 9),
                         stringsAsFactors = FALSE)
 
   expect_equal(as.data.frame(element_summarise_backlog(element, elementid)), # need to coerce result into data frame as expect equal tolerances don't work with tibbles.
@@ -87,9 +93,10 @@ test_that("Summarises backlog by block",{
                         gifa = 1,
                         unit_area = c(1, 1, 2, 10, 20, 30))
 
-  correct <- data.frame(grade = c("B", "C", "B", "C", "D", "E"),
-                        buildingid = c(1, 1, 2, 2, 2, 2),
-                        backlog = c(1, 6, 14, 24, 21.68, 149),
+  correct <- data.frame(grade = rep(c("B", "C", "D", "E"), 2),
+                        buildingid = rep(1:2, each = 4),
+                        backlog = c(1, 6, 0, 0,
+                                    14, 24, 21.68, 149),
                         stringsAsFactors = FALSE)
 
   expect_equal(as.data.frame(element_summarise_backlog(element, buildingid)), # need to coerce result into data frame as expect equal tolerances don't work with tibbles.
@@ -119,9 +126,10 @@ test_that("Summarises area by block",{
                         gifa = 1,
                         unit_area = c(1, 1, 2, 10, 20, 30))
 
-  correct <- data.frame(grade = c("A", "B", "C", "A", "B", "C", "D", "E"),
-                        buildingid = c(1, 1, 1, 2, 2, 2, 2, 2),
-                        area = c(1, 1, 2, 12, 14, 6, 7.1, 20.9),
+  correct <- data.frame(grade = rep(c("A", "B", "C", "D", "E"), 2),
+                        buildingid = rep(1:2, each = 5),
+                        area = c(1, 1, 2, 0, 0,
+                                 12, 14, 6, 7.1, 20.9),
                         stringsAsFactors = FALSE)
 
   expect_equal(as.data.frame(element_summarise_area(element, buildingid)), # need to coerce result into data frame as expect equal tolerances don't work with tibbles.